+++ templates/part.navigation.addfeed.php
@@ -8,53 +8,58 @@
     <div class="add-new-popup" id="new-feed">
 
         <form ng-submit="Navigation.createFeed(feed)" name="feedform">
-            <input type="text"
-                ng-model="feed.url"
-                ng-class="{'ng-invalid': Navigation.feedUrlExists(feed.url)}"
-                placeholder="<?php p($l->t('Web address')); ?>"
-                name="address"
-                pattern="[^\s]+"
-                required>
+            <fieldset ng-disabled="Navigation.addingFeed">
+                <input type="text"
+                    ng-model="feed.url"
+                    ng-class="{'ng-invalid': !Navigation.addingFeed && Navigation.feedUrlExists(feed.url)}"
+                    placeholder="<?php p($l->t('Web address')); ?>"
+                    name="address"
+                    pattern="[^\s]+"
+                    required>
 
-            <p class="error" ng-show="Navigation.feedUrlExists(feed.url)"><?php p($l->t('Feed exists already!')); ?></p>
+                <p class="error"
+                    ng-show="!Navigation.addingFeed && Navigation.feedUrlExists(feed.url)">
+                    <?php p($l->t('Feed exists already!')); ?>
+                </p>
 
-            <!-- select a folder -->
-            <select name="folder"
-                title="<?php p($l->t('Folder')); ?>"
-                ng-if="!Navigation.newFolder"
-                ng-model="$parent.feed.folderId"
-                ng-options="folder.name for folder in Navigation.getFolders() track by folder.name">
-                <option value="">-- <?php p($l->t('No folder')); ?> --</option>
-            </select>
-            <button type="button"
-                    class="icon-add add-new-folder-primary"
-                    ng-hide="Navigation.newFolder"
-                    title="<?php p($l->t('New folder')); ?>"
-                    ng-click="Navigation.newFolder=true"
-                    news-focus="#new-feed [name='folderName']"></button>
+                <!-- select a folder -->
+                <select name="folder"
+                    title="<?php p($l->t('Folder')); ?>"
+                    ng-if="!Navigation.newFolder"
+                    ng-model="$parent.feed.existingFolder"
+                    ng-options="folder.name for folder in Navigation.getFolders() track by folder.name">
+                    <option value="">-- <?php p($l->t('No folder')); ?> --</option>
+                </select>
+                <button type="button"
+                        class="icon-add add-new-folder-primary"
+                        ng-hide="Navigation.newFolder"
+                        title="<?php p($l->t('New folder')); ?>"
+                        ng-click="Navigation.newFolder=true"
+                        news-focus="#new-feed [name='folderName']"></button>
 
-            <!-- add a folder -->
-            <input type="text"
-                   ng-model="$parent.feed.folder"
-                   ng-class="{'ng-invalid': Navigation.newFolder && Navigation.folderNameExists($parent.feed.folder)}"
-                   placeholder="<?php p($l->t('Folder name')); ?>"
-                   name="folderName"
-                   ng-if="Navigation.newFolder"
-                   required>
-            <button type="button"
-                    ng-show="Navigation.newFolder"
-                    class="icon-close add-new-folder-primary"
-                    title="<?php p($l->t('Go back')); ?>"
-                    ng-click="Navigation.newFolder=false; feed.folder=''"></button>
+                <!-- add a folder -->
+                <input type="text"
+                       ng-model="$parent.feed.newFolder"
+                       ng-class="{'ng-invalid': !Navigation.addingFeed && !Navigation.addingFeed && Navigation.newFolder && Navigation.folderNameExists($parent.feed.newFolder)}"
+                       placeholder="<?php p($l->t('Folder name')); ?>"
+                       name="folderName"
+                       ng-if="Navigation.newFolder"
+                       required>
+                <button type="button"
+                        ng-show="Navigation.newFolder"
+                        class="icon-close add-new-folder-primary"
+                        title="<?php p($l->t('Go back')); ?>"
+                        ng-click="Navigation.newFolder=false; feed.folder=''"></button>
 
 
-            <p class="error" ng-show="Navigation.folderNameExists(feed.folder)"><?php p($l->t('Folder exists already!')); ?></p>
+                <p class="error" ng-show="!Navigation.addingFeed && Navigation.folderNameExists(feed.newFolder)"><?php p($l->t('Folder exists already!')); ?></p>
 
-            <input type="submit"
-                value="<?php p($l->t('Subscribe')); ?>"
-                class="primary"
-                ng-disabled="Navigation.feedUrlExists(feed.url) ||
-                            (Navigation.newFolder && Navigation.folderNameExists(folder.name))">
+                <input type="submit"
+                    value="<?php p($l->t('Subscribe')); ?>"
+                    class="primary"
+                    ng-disabled="Navigation.feedUrlExists(feed.url) ||
+                                (Navigation.newFolder && Navigation.folderNameExists(folder.name))">
+            </fieldset>
         </form>
     </div>
 </li>